requirements.txt
aiohttp>=3.8.0
PyYAML>=6.0
numpy>=1.24.0
asyncio>=3.4.3
python-dotenv>=1.0.0